Fixes that allow a client to join a network game
Started work on debugging the network client. This commit fixes
the most glaring problems with client connect: data message
sending and player spawning. Work continues on clientside player
controls, which need to be integrated to the binding/control
system.

+++ doomsday/engine/portable/include/cl_player.h
@@ -34,8 +34,8 @@
 typedef struct clplayerstate_s {
     clmobj_t       *cmo;
     thid_t          mobjId;
-    int             forwardMove;
-    int             sideMove;
+    float           forwardMove;
+    float           sideMove;
     int             angle;
     angle_t         turnDelta;
     int             friction;
